AudioMixer: fill 0s at right place when no more buffers available
When mixer format is PCM_FLOAT and only a partial buffer available
at the end of playback, 0s can be filled at the beginning of the
mixer buffer which will overwrite the copied buffers, then glitch
is heard.

This is because OutPointer is not updated and the pointer always keeps
the start position of mixer buffer when mixer format is PCM_FLOAT.

Assign the value of 'fout' to 'out' to make sure 'out' holds the
current copy position.

CRs-Fixed: 872775

Change-Id: I50504c5a02eb0c69abfc9b047792b0f6f85b9ce8
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>